requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
compress-pickle==2.1.0
croniter==1.1.0
Deprecated==1.2.13
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-RQ2==18.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
imbalanced-learn==0.8.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
psycopg2==2.9.3
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.0
requests==2.26.0
rq==1.10.1
rq-scheduler==0.11.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.29
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3